Refactor the organization of components
